Move algorithms/visit.hpp to algorithms/detail/visit.hpp
